backend/requirements.txt
flask
flask-cors
python-dotenv
requests
numpy
pandas
scikit-learn
